backend/mqtt-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
paho-mqtt==1.6.1
pydantic==2.5.0
